WIP added designToken and renaming
