+++ pwa/src/templates/templateParts/componentCategoryAccordion/ComponentCategoryAccordionTemplate.module.css
