+++ pwa/src/templates/templateParts/layerAccordion/LayerAccordionTemplate.module.css
@@ -1,3 +1,3 @@
-.container {
-  opacity: 0.65;
+.containerInactive {
+  opacity: var(--web-app-layer-accordion-inactive-opacity);
 }
