+++ pwa/src/templates/templateParts/layerAccordion/header/LayerAccordionHeaderTemplate.tsx
@@ -3,10 +3,11 @@ import * as styles from "./LayerAccordionHeaderTemplate.module.css";
 import clsx from "clsx";
 import { Divider, Heading3 } from "@gemeente-denhaag/components-react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faChevronRight, faLayerGroup } from "@fortawesome/free-solid-svg-icons";
+import { faChevronRight, faLayerGroup, faRepeat } from "@fortawesome/free-solid-svg-icons";
 import { useTranslation } from "react-i18next";
 import { BadgeCounter } from "../../../../components/badgeCounter/BadgeCounter";
 import _ from "lodash";
+import { Tag } from "../../../../components/tag/Tag";
 
 interface LayerAccordionHeaderTemplateProps {
   title: string;
@@ -30,9 +31,10 @@ export const LayerAccordionHeaderTemplate: React.FC<LayerAccordionHeaderTemplate
             <FontAwesomeIcon className={styles.layerIcon} icon={faLayerGroup} />
             {badgeNumber > 0 && (
               <>
-                <BadgeCounter number={badgeNumber} stylingClassName={styles[_.camelCase(`${title} badge`)]}>
-                  <Heading3 className={styles.title}>{t(title)}</Heading3>
-                </BadgeCounter>
+                {/* <BadgeCounter number={badgeNumber} stylingClassName={styles[_.camelCase(`${title} badge`)]}> */}
+                <Heading3 className={styles.title}>{t(title)}</Heading3>
+                {/* </BadgeCounter> */}
+                <Tag label={badgeNumber.toString()} icon={<FontAwesomeIcon icon={faRepeat} />} />
                 <FontAwesomeIcon className={clsx(styles.toggleIcon, open && styles.open)} icon={faChevronRight} />
               </>
             )}
