+++ pwa/src/templates/templateParts/resultsTemplates/layers/LayersResultTemplate.tsx
@@ -2,12 +2,12 @@ import * as React from "react";
 import * as styles from "./LayersResultTemplate.module.css";
 
 import _ from "lodash";
-import { CardsSearchTemplate } from "../../CardsSearch/CardsSearchTemplate";
+import { ComponentCardsAccordionTemplate } from "../../componentCardsAccordion/ComponentCardsAccordionTemplate";
 
 interface LayersResultTemplateProps {
   components: any[];
 }
 
 export const LayersResultTemplate: React.FC<LayersResultTemplateProps> = ({ components }) => {
-  return <CardsSearchTemplate components={components} />;
+  return <ComponentCardsAccordionTemplate components={components} />;
 };
